+++ src/Main.hs
@@ -17,4 +17,5 @@ main = do
   let expr = genExpr exprString
   putStrLn (show $ expr)
   putStrLn (exprToString expr)
+  putStrLn (show $ countVars expr)
   putStrLn (exprToString $ simplifyProducts expr)